Last commit

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,22 +1,23 @@
 #!/usr/bin/python3
 """ Module for task 13 """
 
-import sys
-from model_state import Base, State
 from sqlalchemy import create_engine
-from sqlalchemy.orm import Session
-from sqlalchemy.schema import Table
+from sqlalchemy.orm import sessionmaker
+from model_state import State
+import sys
 
 
 if __name__ == "__main__":
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost:3306/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
+    engine.connect()
+    Session = sessionmaker(bind=engine)
+    session = Session()
 
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
-        sys.argv[1], sys.argv[2],
-        sys.argv[3]), pool_pre_ping=True)
-    Base.metadata.create_all(engine)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
+    session.commit()
 
-    session = Session(engine)
-    for state in session.query(State).filter(State.name.like('%a%')):
-        session.delete(state)
-        session.commit()
-        session.close()
+    session.close()
